# back_end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recherche d'emploi",
        "description": "Documentation de l'API Laravel pour la plateforme de recherche d'emploi",
        "contact": {
            "email": "hariznabil663@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/personal-data": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get personal data of the authenticated user",
                "operationId": "71f884209dd7c91ed10dc2941cd009ca",
                "responses": {
                    "200": {
                        "description": "User data returned successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Submit a job application",
                "operationId": "ca31bd612c53c5417427c27da3dc99ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "job_id",
                                    "phone_number",
                                    "cover_letter",
                                    "resume_path"
                                ],
                                "properties": {
                                    "job_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "123-456-7890"
                                    },
                                    "cover_letter": {
                                        "type": "string",
                                        "example": "I'm very interested in this role..."
                                    },
                                    "resume_path": {
                                        "type": "string",
                                        "example": "/storage/resumes/user123.pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/applications": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get applications submitted by the authenticated user",
                "operationId": "a4586725dbaac691eddd0ffe0e532e1e",
                "responses": {
                    "200": {
                        "description": "List of user's job applications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "applications": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "job_id": {
                                                        "type": "integer"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}